requirements.txt
gradio>=4.44.0,<6
ollama>=0.4.7,<1
mlx-whisper>=0.4.0
pyannote.audio==3.1.1
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
transformers>=4.39.0,<5
huggingface_hub>=0.22.0,<1
